refactor: 外移 LLM/Common 超长文件的内联测试模块
将 qq-maid-common/src/time_context.rs 与 qq-maid-llm/src/provider/mod.rs
尾部的 #[cfg(test)] mod tests 整体外移：
- time_context.rs 拆为 time_context/mod.rs + time_context/tests.rs 目录布局
- provider/mod.rs 末尾改 #[cfg(test)] mod tests; 并新增同级 provider/tests.rs

仅机械移动完整测试模块，保留测试注释与断言语义，不改业务逻辑、
公开 API、provider 协议、fallback、SSE、usage 或时间/时区语义。

refs #110
